+++ Scryber.Components.Mvc/ControllerExtensions.cs
@@ -99,6 +99,7 @@ namespace Scryber.Components.Mvc
 
         #endregion
 
+
         #region Generic type model
 
         /// <summary>
@@ -189,6 +190,90 @@ namespace Scryber.Components.Mvc
         }
 
         #endregion
+
+        //
+        // Async requests
+        //
+
+        /// <summary>
+        /// Generate and return a PDF document file result from the template at the specified document path
+        /// </summary>
+        /// <param name="controller">Self reference to the controller as an extension method</param>
+        /// <param name="fullpath">The full path to the document in the file system</param>
+        /// <param name="inline">If true, then the output PDF will be directly streamed to the output, otherwise it will be as an attachment</param>
+        /// <param name="outputFileName">The name of the file to be downloaded if it is an attachment, if not set the document file name or controller request action name will be used</param>
+        /// <returns>A file action result with the appropriate headers and content type</returns>
+        public static IActionResult PDF(this ControllerBase controller, string fullpath, bool inline = true, string outputFileName = "")
+        {
+            var doc = Document.ParseDocument(fullpath);
+            if (inline == false)
+            {
+                if (string.IsNullOrEmpty(outputFileName))
+                    outputFileName = System.IO.Path.GetFileNameWithoutExtension(fullpath) + ".pdf";
+            }
+
+            return PDF(controller, doc, inline, outputFileName);
+        }
+
+        /// <summary>
+        /// Generate and return a PDF document file result from the template in the content stream
+        /// </summary>
+        /// <param name="controller">Self reference to the controller as an extension method</param>
+        /// <param name="content">The xml template content within the stream</param>
+        /// <param name="inline">If true, then the output PDF will be directly streamed to the output, otherwise it will be as an attachment</param>
+        /// <param name="outputFileName">The name of the file to be downloaded if it is an attachment, if not set the document file name or controller request action name will be used</param>
+        /// <returns>A file action result with the appropriate headers and content type</returns>
+        public static async Task<IActionResult> PDFAsync(this ControllerBase controller, System.IO.Stream content, bool inline = true, string outputFileName = "")
+        {
+            var doc = Document.ParseDocument(content, ParseSourceType.DynamicContent);
+
+            return await PDFAsync(controller, doc, inline, outputFileName);
+        }
+
+        /// <summary>
+        /// Generate and return a PDF document file result from the template in the content stream
+        /// </summary>
+        /// <param name="controller">Self reference to the controller as an extension method</param>
+        /// <param name="content">The xml template content within the stream</param>
+        /// <param name="path">The path to the file for any relative references</param>
+        /// <param name="inline">If true, then the output PDF will be directly streamed to the output, otherwise it will be as an attachment</param>
+        /// <param name="outputFileName">The name of the file to be downloaded if it is an attachment, if not set the document file name or controller request action name will be used</param>
+        /// <returns>A file action result with the appropriate headers and content type</returns>
+        public static async Task<IActionResult> PDFAsync(this ControllerBase controller, System.IO.Stream content, string path, bool inline = true, string outputFileName = "")
+        {
+            var doc = Document.ParseDocument(content, path, ParseSourceType.DynamicContent);
+
+            return PDF(controller, doc, inline, outputFileName);
+        }
+
+        /// <summary>
+        /// Generate and return a PDF document file result from the PDFDocument instance
+        /// </summary>
+        /// <param name="controller">Self reference to the controller as an extension method</param>
+        /// <param name="document">The PDF document to render to the output stream</param>
+        /// <param name="inline">If true, then the output PDF will be directly streamed to the output, otherwise it will be as an attachment</param>
+        /// <param name="outputFileName">The name of the file to be downloaded if it is an attachment, if not set the document file name or controller request action name will be used</param>
+        /// <returns>A file action result with the appropriate headers and content type</returns>
+        public static async Task<IActionResult> PDFAsync(this ControllerBase controller, Document document, bool inline = true, string outputFileName = "")
+        {
+            var ms = new System.IO.MemoryStream();
+
+            await document.SaveAsPDFAsync(ms);
+
+            ms.Flush();
+
+            ms.Position = 0;
+            if (!inline)
+            {
+                if (string.IsNullOrEmpty(outputFileName))
+                    outputFileName = GetDocumentName(document, controller);
+                return controller.File(ms, PDFMimeType, outputFileName);
+            }
+            else
+                return controller.File(ms, PDFMimeType);
+        }
+
+
 
         #region ParseView[<TModel>](this Controller controller...)